# apps/idx-api/openapi.yaml
openapi: "3.0.3"
info:
  title: Idx Auth API
  description: |
    Authentication, authorization, user management, and chat thread API for Idx Chat.
    
    ## Response Envelope
    Every response follows a standardized format:
    ```json
    {
      "success": true | false,
      "requestId": "uuid-v4",
      "data": { ... },
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message",
        "details": { ... }
      }
    }
    ```
    
    ## Authentication
    Most endpoints require a `Bearer` token in the `Authorization` header.
    Obtain tokens via `/auth/login`, `/auth/register`, or `/auth/google`.
  version: "0.1.0"
  contact:
    name: Idx Team
servers:
  - url: http://localhost:4000
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required: [success, requestId]
      properties:
        success: { type: boolean }
        requestId: { type: string, format: uuid }
        data: {}
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
...
    UserProfile:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserProfile"
        - type: object
          properties:
            status: { type: string }
            createdAt: { type: string, format: date-time }
...
    RoleDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
    PermissionDetail:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        resource: { type: string }
        action: { type: string }
...
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items: { $ref: "#/components/schemas/UserDetail" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get user details with roles & permissions
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User detail
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: Update user display name
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
      responses:
        "200":
          description: User updated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin]
      summary: Delete user account
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/roles:
    post:
      tags: [Admin]
      summary: Assign role to user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleName]
              properties:
                roleName: { type: string }
      responses:
        "200":
          description: Role assigned
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Admin]
      summary: Revoke role from user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleName]
              properties:
                roleName: { type: string }
      responses:
        "200":
          description: Role revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/ban:
    patch:
      tags: [Admin]
      summary: Ban or unban a user
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, banned] }
      responses:
        "200":
          description: Status updated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/reset-password:
    post:
      tags: [Admin]
      summary: Force reset user password
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password reset
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/force-logout:
    post:
      tags: [Admin]
      summary: Revoke all user sessions
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Sessions revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/roles:
    get:
      tags: [Admin]
      summary: List all roles
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          roles:
                            type: array
                            items: { $ref: "#/components/schemas/RoleDetail" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/roles/{id}/permissions:
    get:
      tags: [Admin]
      summary: List permissions for a role
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Permission list
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Assign permission to role
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissionId]
              properties:
                permissionId: { type: integer }
      responses:
        "200":
          description: Permission assigned
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Admin]
      summary: Revoke permission from role
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissionId]
              properties:
                permissionId: { type: integer }
      responses:
        "200":
          description: Permission revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/permissions:
    get:
      tags: [Admin]
      summary: List all permissions
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Permission list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          permissions:
                            type: array
                            items: { $ref: "#/components/schemas/PermissionDetail" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/stats:
    get:
      tags: [Admin]
      summary: System statistics
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userCount: { type: integer }
                          roleCount: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Threads ─────────────────────────────────────────
...